requirements.txt
Django==4.2.0
djangorestframework==3.14.0
python-crontab==3.0.0
django-crontab==0.7.1
